requirements.txt
# Database & ORM
SQLAlchemy==2.0.36
Alembic==1.17.0
psycopg2-binary==2.9.10

# Web Framework
Flask==3.1.1

# Article Discovery
feedparser==6.0.11
anthropic==0.40.0
exa-py==1.1.0
httpx==0.28.1

# Email Delivery
sendgrid==6.11.0

# Deep Dive / Google Integration
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
beautifulsoup4==4.12.3

# Testing
pytest==8.3.4
pytest-postgresql==6.1.1

# Server
gunicorn==23.0.0

# Utilities
python-dotenv==1.0.1

